fix: First interaction leads to errors and breaks the app
Now checks on each request if the user.id exists else it creates it and then sends the request.

Had to fix the backend middleware as it was not responding to my registration requests, and that was because the middleware did not run next(); if the URL matched the registration endpoint.

Also removed the code block that checks if the user is registered from the MainLayout. It's not necessary anymore.